test: add missing unit tests for NativeHolderSmartWallet
